Many changes
* Convert the battleship firing mode to editing a message, rather than
  sending messages.
* Add message when ship is sunk.
* Add a "help" argument to the pm_games _game functions. If provided,
  this will add a button to send help for that game in a DM.
* Updated i18n files to reflect that change.
* Converted @client.event on_message to @client.listen().
* Reworked delete handlers to use @client.listen() and
  @client.remove_listener().
* Removed blackjack for good as it took up space and was badly made.
* Changed i18n on_raw_reaction_add to use the and keyword which
  shortcuts, rather than all(()) which doesn't.

+++ kenny2automate/__main__.py
@@ -62,9 +62,7 @@ client = Bot(
 	activity=d.Activity(type=d.ActivityType.watching, name=';help')
 )
 
-DGDELETEHANDLERS = {}
-
-@client.event
+@client.listen()
 async def on_message(msg):
 	ctx = await client.get_context(msg)
 	if (ctx.guild
@@ -75,18 +73,11 @@ async def on_message(msg):
 				msg.content, re.I):
 			await msg.delete()
 			return
-	await client.invoke(ctx)
 
 @client.event
 async def on_message_delete(msg):
 	if msg.mentions:
 		logger.info('Message with mentions deleted: {}'.format(msg.content), extra={'ctx': DummyCtx(author=DummyCtx(name=msg.author.name))})
-	if msg.id in DGDELETEHANDLERS:
-		await DGDELETEHANDLERS[msg.id](msg)
-		del DGDELETEHANDLERS[msg.id]
-
-def add_delete_handler(msg, handler):
-	DGDELETEHANDLERS[msg.id] = handler
 
 @client.event
 async def on_command_error(ctx, exc):
@@ -154,9 +145,11 @@ async def repeat(ctx, *, arg):
 	"""Repeat what you say, right back at ya."""
 	logger.info('repeat: ' + arg, extra={'ctx': ctx})
 	msg = await ctx.send(arg)
-	async def handle_delete(m):
-		await msg.delete()
-	add_delete_handler(ctx.message, handle_delete)
+	@client.listen()
+	async def on_message_delete(msg2):
+		if msg2.id == msg.id:
+			await msg.delete()
+			client.remove_listener(on_message_delete)
 
 @client.command()
 async def hello(ctx):
